authentication and redirect working
